+++ tests/test_cli_import_solution.py
@@ -651,12 +651,12 @@ class TestCliImportSolution(object):  # pylint: disable=too-many-public-methods
         be the same for all content types.
         """
 
-        file_content = mock.mock_open(read_data=Const.NEWLINE.join(Solution.TEMPLATE))
+        file_content = mock.mock_open(read_data=Const.NEWLINE.join(Solution.TEMPLATE_TEXT))
         with mock.patch('snippy.content.migrate.open', file_content, create=True) as mock_file:
-            cause = snippy.run(['snippy', 'import', '--solution', '--template'])
+            cause = snippy.run(['snippy', 'import', '--solution', '--template', '--format', 'text'])
             assert cause == 'NOK: content was not stored because it was matching to an empty template'
             Content.assert_storage(None)
-            mock_file.assert_called_once_with('./solution-template.txt', 'r')
+            mock_file.assert_called_once_with('./solution-template.text', 'r')
 
     @pytest.mark.usefixtures('yaml')
     def test_cli_import_solution_029(self, snippy):
@@ -731,6 +731,22 @@ class TestCliImportSolution(object):  # pylint: disable=too-many-public-methods
             assert cause == Cause.ALL_OK
             Content.assert_storage(content)
             mock_file.assert_called_once_with('./all-solutions.md', 'r')
+
+    @pytest.mark.usefixtures('isfile_true')
+    def test_cli_import_solution_033(self, snippy):
+        """Import solutions from Markdown template.
+
+        Try to import solution template without any changes. This should result
+        error text for end user and no files should be read. The error text must
+        be the same for all content types.
+        """
+
+        file_content = mock.mock_open(read_data=Const.NEWLINE.join(Solution.TEMPLATE_MKDN))
+        with mock.patch('snippy.content.migrate.open', file_content, create=True) as mock_file:
+            cause = snippy.run(['snippy', 'import', '--solution', '--template', '--format', 'mkdn'])
+            assert cause == 'NOK: content was not stored because it was matching to an empty template'
+            Content.assert_storage(None)
+            mock_file.assert_called_once_with('./solution-template.mkdn', 'r')
 
     @classmethod
     def teardown_class(cls):