+++ tests/testlib/reference.py
@@ -108,5 +108,45 @@ class Reference(object):  # pylint: disable=too-few-public-methods
     DEFAULT_REFERENCES = (GITLOG, REGEXP)
 
     TEMPLATE = Helper.read_template('reference.txt').split('\n')
-    TEMPLATE_DIGEST_TEXT = 'e0cd55c650ef936a66633ee29500e47ee60cc497c342212381c40032ea2850d9'
-    TEMPLATE_DIGEST_MKDN = '5f8abf17149b097453571524f403754997682cdbcfc6de8605bb0eff19fc507e'
+    TEMPLATE_DIGEST_TEXT_MODEL = 'e0cd55c650ef936a66633ee29500e47ee60cc497c342212381c40032ea2850d9'
+    TEMPLATE_DIGEST_MKDN_MODEL = '5f8abf17149b097453571524f403754997682cdbcfc6de8605bb0eff19fc507e'
+    TEMPLATE_TEXT = (
+        '# Commented lines will be ignored.',
+        '#',
+        '# Add mandatory links below one link per line.',
+        'https://www.example.com/add-links-here.html',
+        '',
+        '# Add optional brief description below.',
+        'Add brief title for content',
+        '',
+        '# Add optional description below.',
+        'Add a description that defines the content in one chapter.',
+        '',
+        '# Add optional comma separated list of groups below.',
+        'groups',
+        '',
+        '# Add optional comma separated list of tags below.',
+        'comma,separated,tags',
+        ''
+    )
+    TEMPLATE_MKDN = (
+        '# Add brief title for content @groups',
+        '',
+        '> Add a description that defines the content in one chapter.',
+        '',
+        '> [1] https://www.example.com/add-links-here.html',
+        '',
+        '## Meta',
+        '',
+        '> category : reference  ',
+        'created  : 2018-02-02T02:02:02.000001+00:00  ',
+        'digest   : 5f8abf17149b097453571524f403754997682cdbcfc6de8605bb0eff19fc507e  ',
+        'filename :   ',
+        'name     :   ',
+        'source   :   ',
+        'tags     : comma,separated,tags  ',
+        'updated  : 2018-02-02T02:02:02.000001+00:00  ',
+        'uuid     : 1acd5827-b6ef-4067-b5ac-3ceac07dde9f  ',
+        'versions : ',
+        ''
+    )